refactor: 폴더 이동, 필터버튼 스타일 적용
